requirements.txt
aniso8601==8.0.0
boto3==1.9.252
botocore==1.12.252
certifi==2019.9.11
chardet==3.0.4
Click==7.0
docutils==0.15.2
Flask==1.1.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
passlib==1.7.1
psycopg2==2.8.3
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
s3transfer==0.2.1
six==1.12.0
SQLAlchemy==1.3.10
stripe==2.37.2
urllib3==1.25.6
Werkzeug==0.16.0
